Added headers for vocab, vectors, and regex tokenizers
