+++ torchtext/csrc/common.h
@@ -1,0 +1,3 @@
+namespace torchtext {
+typedef std::vector<std::string> StringList;
+} // namespace torchtext
