+++ torchtext/csrc/regex_tokenizer.cpp
@@ -44,30 +44,4 @@ void RegexTokenizer::split_(std::string &str, std::vector<std::string> &tokens,
   }
 }
 
-// Registers our custom class with torch.
-static auto regex_tokenizer =
-    torch::class_<RegexTokenizer>("torchtext", "RegexTokenizer")
-        .def(torch::init<std::vector<std::string>, std::vector<std::string>,
-                         bool>())
-        .def("forward", &RegexTokenizer::forward)
-        .def_pickle(
-            // __setstate__
-            [](const c10::intrusive_ptr<RegexTokenizer> &self)
-                -> std::tuple<std::vector<std::string>,
-                              std::vector<std::string>, bool> {
-              return std::make_tuple(self->patterns_, self->replacements_,
-                                     self->to_lower_);
-            },
-            // __getstate__
-            [](std::tuple<std::vector<std::string>, std::vector<std::string>,
-                          bool>
-                   states) -> c10::intrusive_ptr<RegexTokenizer> {
-              auto patterns = std::get<0>(states);
-              auto replacements = std::get<1>(states);
-              auto to_lower = std::get<2>(states);
-
-              return c10::make_intrusive<RegexTokenizer>(
-                  std::move(patterns), std::move(replacements), to_lower);
-            });
-
 } // namespace torchtext
